IO Writing optimizations
Moved writing of bits to the wire to be
immediate by every writer rather than by the general
I/O thread, thus drastically improving responsiveness.
Also moved the main OFMessageAsyncStream into our code
class. As a result of other threads handling the writes
it was necessary to provide a method to early detect
when a remote switch disconnected preventing writes,
thus write exceptions are now swollowed, but the failure
to write is detected by the end of the full handling of
a single message and the switch properly disconnected by
the I/O thread.