Fixed malformed OpenFlow message crashing worker thread
A failure during decoding of an OpenFlow message could cause
a RuntimeException to be thrown which was only caught by
the parent worker thread, which as a result would kill the
thread, and any switches attached to it, or attached in
the future would be unable to maintain a connection.  The
solution is to convert all Exceptions to IOException from
within the OFMessageInStream so we can safely catch them
as IOExceptions in the Controller, disconnect the switch,
and log the problem.

Thanks to Ali Khalfan for the bug report.